requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
node==1.1
nodejs==0.1.1
npm==0.1.1
odict==1.9.0
optional-django==0.1.0
plumber==1.7
pytz==2022.6
sqlparse==0.4.3
zope.component==5.0.1
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.3
zope.interface==5.5.1
zope.lifecycleevent==4.4
zope.proxy==4.6.0
